move content of system_bar to dev page
